Remove bad DNSBL ips.backscatterer.org
The DNSBL ips.backscatterer.org is annoying and blacklists people for
mostly bogus reasons and actually charges money for delisting. I'm
never going to pay money to be removed from this DNSBL, and I'm also
not going to change the things that this DNSBL thinks I should change,
so there's no point in checking it.