Don't check polspam.pl
Polspam.pl does not provide a tool on their site for checking with
lists someone is in, and the delisting procedure they describe on
their site is convoluted, confusing, and hostile, and unlikely to be
effectively usable. Furthermore, they claim that sites are removed
automatically when spamming stops. It doesn't seem to worthwhile to be
checking them.